+++ lib/io_client.dart
@@ -1,2 +1,3 @@
 export 'src/base.dart';
 export 'src/version.dart';
+export 'src/io.dart';