+++ lib/sentry_browser.dart
@@ -1,6 +1,0 @@
-library sentry.browser;
-
-export 'src/base.dart';
-export 'src/stack_trace.dart';
-export 'src/browser.dart';
-export 'src/version.dart';