+++ test/sentry_browser_test.dart
@@ -4,48 +4,78 @@
 
 @TestOn("browser")
 import 'dart:convert';
-import 'dart:html' hide Event;
 
 import 'package:http/http.dart';
-import 'package:sentry/sentry_browser.dart';
+import 'package:sentry/browser_client.dart';
 import 'package:test/test.dart';
-
-import 'test_utils.dart';
 
 const String _testDsn = 'https://public:secret@sentry.example.com/1';
 const String _testDsnWithoutSecret = 'https://public@sentry.example.com/1';
-
+const String _testDsnWithPath =
+    'https://public:secret@sentry.example.com/path/1';
+const String _testDsnWithPort =
+    'https://public:secret@sentry.example.com:8888/1';
 void main() {
-  group('$SentryClientBrowser', () {
+  group('$SentryBrowserClient', () {
     test('can parse DSN', () async {
-      final SentryClientBase client = new SentryClientBrowser(dsn: _testDsn);
-      testDsn(client, _testDsn);
+      final SentryBrowserClient client = SentryBrowserClient(dsn: _testDsn);
+      expect(client.dsnUri, Uri.parse(_testDsn));
+      expect(client.postUri, 'https://sentry.example.com/api/1/store/');
+      expect(client.publicKey, 'public');
+      expect(client.secretKey, 'secret');
+      expect(client.projectId, '1');
       await client.close();
     });
 
     test('can parse DSN without secret', () async {
-      final SentryClientBase client =
-          new SentryClientBrowser(dsn: _testDsnWithoutSecret);
-      testDsn(client, _testDsnWithoutSecret, withSecret: false);
+      final SentryBrowserClient client =
+          SentryBrowserClient(dsn: _testDsnWithoutSecret);
+      expect(client.dsnUri, Uri.parse(_testDsnWithoutSecret));
+      expect(client.postUri, 'https://sentry.example.com/api/1/store/');
+      expect(client.publicKey, 'public');
+      expect(client.secretKey, null);
+      expect(client.projectId, '1');
       await client.close();
     });
 
-    test('sends client auth header without secret and without user agent',
-        () async {
-      final ClockProvider fakeClockProvider =
-          () => new DateTime.utc(2017, 1, 2);
+    test('can parse DSN with path', () async {
+      final SentryBrowserClient client =
+          SentryBrowserClient(dsn: _testDsnWithPath);
+      expect(client.dsnUri, Uri.parse(_testDsnWithPath));
+      expect(client.postUri, 'https://sentry.example.com/path/api/1/store/');
+      expect(client.publicKey, 'public');
+      expect(client.secretKey, 'secret');
+      expect(client.projectId, '1');
+      await client.close();
+    });
+    test('can parse DSN with port', () async {
+      final SentryBrowserClient client =
+          SentryBrowserClient(dsn: _testDsnWithPort);
+      expect(client.dsnUri, Uri.parse(_testDsnWithPort));
+      expect(client.postUri, 'https://sentry.example.com:8888/api/1/store/');
+      expect(client.publicKey, 'public');
+      expect(client.secretKey, 'secret');
+      expect(client.projectId, '1');
+      await client.close();
+    });
+    test('sends client auth header without secret', () async {
+      final MockClient httpMock = MockClient();
+      final ClockProvider fakeClockProvider = () => DateTime.utc(2017, 1, 2);
 
       Map<String, String> headers;
 
-      final MockClient httpMock = new MockClient((Request request) async {
-        if (request.method == "POST") {
-          headers = request.headers;
-          return new Response('{"id": "test-event-id"}', 200);
+      httpMock.answerWith((Invocation invocation) async {
+        if (invocation.memberName == #close) {
+          return null;
         }
-        return new Response('Unexpected invocation in HttpMock', 404);
+        if (invocation.memberName == #post) {
+          headers = invocation.namedArguments[#headers];
+          return Response('{"id": "test-event-id"}', 200);
+        }
+        fail('Unexpected invocation of ${invocation.memberName} in HttpMock');
       });
 
-      final SentryClientBase client = new SentryClientBrowser(
+      final SentryBrowserClient client = SentryBrowserClient(
         dsn: _testDsnWithoutSecret,
         httpClient: httpMock,
         clock: fakeClockProvider,
@@ -58,7 +88,7 @@ void main() {
       );
 
       try {
-        throw new ArgumentError('Test error');
+        throw ArgumentError('Test error');
       } catch (error, stackTrace) {
         final SentryResponse response = await client.captureException(
             exception: error, stackTrace: stackTrace);
@@ -67,35 +97,40 @@ void main() {
         expect(response.error, null);
       }
 
-      testHeaders(
-        headers,
-        fakeClockProvider,
-        withUserAgent: false,
-        compressPayload: false,
-        withSecret: false,
-      );
+      final Map<String, String> expectedHeaders = <String, String>{
+        'Content-Type': 'application/json',
+        'X-Sentry-Auth': 'Sentry sentry_version=6, '
+            'sentry_client=${SentryClient.sentryClient}, '
+            'sentry_timestamp=${fakeClockProvider().millisecondsSinceEpoch}, '
+            'sentry_key=public',
+      };
+
+      expect(headers, expectedHeaders);
 
       await client.close();
     });
 
     testCaptureException() async {
-      final ClockProvider fakeClockProvider =
-          () => new DateTime.utc(2017, 1, 2);
+      final MockClient httpMock = MockClient();
+      final ClockProvider fakeClockProvider = () => DateTime.utc(2017, 1, 2);
 
       String postUri;
       Map<String, String> headers;
       List<int> body;
-      final MockClient httpMock = new MockClient((Request request) async {
-        if (request.method == "POST") {
-          postUri = request.url.toString();
-          headers = request.headers;
-          body = request.bodyBytes;
-          return new Response('{"id": "test-event-id"}', 200);
+      httpMock.answerWith((Invocation invocation) async {
+        if (invocation.memberName == #close) {
+          return null;
         }
-        return new Response('Unexpected invocation in HttpMock', 404);
+        if (invocation.memberName == #post) {
+          postUri = invocation.positionalArguments.single;
+          headers = invocation.namedArguments[#headers];
+          body = invocation.namedArguments[#body];
+          return Response('{"id": "test-event-id"}', 200);
+        }
+        fail('Unexpected invocation of ${invocation.memberName} in HttpMock');
       });
 
-      final SentryClientBase client = new SentryClientBrowser(
+      final SentryBrowserClient client = SentryBrowserClient(
         dsn: _testDsn,
         httpClient: httpMock,
         clock: fakeClockProvider,
@@ -108,7 +143,7 @@ void main() {
       );
 
       try {
-        throw new ArgumentError('Test error');
+        throw ArgumentError('Test error');
       } catch (error, stackTrace) {
         final SentryResponse response = await client.captureException(
             exception: error, stackTrace: stackTrace);
@@ -119,14 +154,19 @@ void main() {
 
       expect(postUri, client.postUri);
 
-      testHeaders(
-        headers,
-        fakeClockProvider,
-        withUserAgent: false,
-        compressPayload: false,
-      );
+      final Map<String, String> expectedHeaders = <String, String>{
+        'Content-Type': 'application/json',
+        'X-Sentry-Auth': 'Sentry sentry_version=6, '
+            'sentry_client=${SentryClient.sentryClient}, '
+            'sentry_timestamp=${fakeClockProvider().millisecondsSinceEpoch}, '
+            'sentry_key=public, '
+            'sentry_secret=secret',
+      };
 
-      final Map<String, dynamic> data = json.decode(utf8.decode(body));
+      expect(headers, expectedHeaders);
+
+      Map<String, dynamic> data = json.decode(utf8.decode(body));
+
       final Map<String, dynamic> stacktrace = data.remove('stacktrace');
       expect(stacktrace['frames'], const TypeMatcher<List>());
       expect(stacktrace['frames'], isNotEmpty);
@@ -139,40 +179,246 @@ void main() {
         'lineno',
         'colno',
         'in_app',
-        'filename'
+        'filename',
       ]);
-      expect(topFrame['abs_path'],
-          '${window.location.origin}/sentry_browser_test.dart.browser_test.dart.js');
-      // fixme: the function name should be 'main.<fn>.testCaptureException'
-      // but it is not due to javascript stacktrace
-      // find a way to correctly test that
+
+      // can't test full url, local PORT can change
+      expect(topFrame['abs_path'].startsWith('http://localhost:'), isTrue);
+      expect(
+        topFrame['abs_path']
+            .endsWith('/sentry_browser_test.dart.browser_test.dart.js'),
+        isTrue,
+      );
       expect(topFrame['function'], 'Object.wrapException');
       expect(topFrame['lineno'], greaterThan(0));
-      expect(topFrame['colno'], greaterThan(0));
       expect(topFrame['in_app'], true);
-      expect(topFrame['filename'],
-          'sentry_browser_test.dart.browser_test.dart.js');
+      expect(topFrame['filename'], 'sentry_browser_test.dart.browser_test.dart.js');
 
       expect(data, {
         'project': '1',
         'event_id': 'X' * 32,
         'timestamp': '2017-01-02T00:00:00',
+        'platform': 'javascript',
         'exception': [
           {'type': 'ArgumentError', 'value': 'Invalid argument(s): Test error'}
         ],
         'sdk': {'version': sdkVersion, 'name': 'dart'},
-        'logger': SentryClientBase.defaultLoggerName,
+        'logger': 'SentryClient',
         'server_name': 'test.server.com',
         'release': '1.2.3',
         'environment': 'staging',
-        'platform': 'javascript',
       });
 
       await client.close();
     }
 
-    test('sends an exception report', () async {
+    test('sends an exception report (uncompressed)', () async {
       await testCaptureException();
     });
+
+    test('reads error message from the x-sentry-error header', () async {
+      final MockClient httpMock = MockClient();
+      final ClockProvider fakeClockProvider = () => DateTime.utc(2017, 1, 2);
+
+      httpMock.answerWith((Invocation invocation) async {
+        if (invocation.memberName == #close) {
+          return null;
+        }
+        if (invocation.memberName == #post) {
+          return Response('', 401, headers: <String, String>{
+            'x-sentry-error': 'Invalid api key',
+          });
+        }
+        fail('Unexpected invocation of ${invocation.memberName} in HttpMock');
+      });
+
+      final SentryBrowserClient client = SentryBrowserClient(
+        dsn: _testDsn,
+        httpClient: httpMock,
+        clock: fakeClockProvider,
+        uuidGenerator: () => 'X' * 32,
+        environmentAttributes: const Event(
+          serverName: 'test.server.com',
+          release: '1.2.3',
+          environment: 'staging',
+        ),
+      );
+
+      try {
+        throw ArgumentError('Test error');
+      } catch (error, stackTrace) {
+        final SentryResponse response = await client.captureException(
+            exception: error, stackTrace: stackTrace);
+        expect(response.isSuccessful, false);
+        expect(response.eventId, null);
+        expect(response.error,
+            'Sentry.io responded with HTTP 401: Invalid api key');
+      }
+
+      await client.close();
+    });
+
+    test('$Event userContext overrides client', () async {
+      final MockClient httpMock = MockClient();
+      final ClockProvider fakeClockProvider = () => DateTime.utc(2017, 1, 2);
+
+      String loggedUserId; // used to find out what user context was sent
+      httpMock.answerWith((Invocation invocation) async {
+        if (invocation.memberName == #close) {
+          return null;
+        }
+        if (invocation.memberName == #post) {
+          // parse the body and detect which user context was sent
+          var bodyData = invocation.namedArguments[Symbol("body")];
+          var decoded = Utf8Codec().decode(bodyData);
+          var decodedJson = JsonDecoder().convert(decoded);
+          loggedUserId = decodedJson['user']['id'];
+          return Response('', 401, headers: <String, String>{
+            'x-sentry-error': 'Invalid api key',
+          });
+        }
+        fail('Unexpected invocation of ${invocation.memberName} in HttpMock');
+      });
+
+      final clientUserContext = User(
+          id: "client_user",
+          username: "username",
+          email: "email@email.com",
+          ipAddress: "127.0.0.1");
+      final eventUserContext = User(
+          id: "event_user",
+          username: "username",
+          email: "email@email.com",
+          ipAddress: "127.0.0.1",
+          extras: {"foo": "bar"});
+
+      final SentryBrowserClient client = SentryBrowserClient(
+        dsn: _testDsn,
+        httpClient: httpMock,
+        clock: fakeClockProvider,
+        uuidGenerator: () => 'X' * 32,
+        environmentAttributes: const Event(
+          serverName: 'test.server.com',
+          release: '1.2.3',
+          environment: 'staging',
+        ),
+      );
+      client.userContext = clientUserContext;
+
+      try {
+        throw ArgumentError('Test error');
+      } catch (error, stackTrace) {
+        final eventWithoutContext =
+            Event(exception: error, stackTrace: stackTrace);
+        final eventWithContext = Event(
+            exception: error,
+            stackTrace: stackTrace,
+            userContext: eventUserContext);
+        await client.capture(event: eventWithoutContext);
+        expect(loggedUserId, clientUserContext.id);
+        await client.capture(event: eventWithContext);
+        expect(loggedUserId, eventUserContext.id);
+      }
+
+      await client.close();
+    });
   });
+
+  group('$Event', () {
+    test('$Breadcrumb serializes', () {
+      expect(
+        Breadcrumb(
+          "example log",
+          DateTime.utc(2019),
+          level: SeverityLevel.debug,
+          category: "test",
+        ).toJson(),
+        <String, dynamic>{
+          'timestamp': '2019-01-01T00:00:00',
+          'message': 'example log',
+          'category': 'test',
+          'level': 'debug',
+        },
+      );
+    });
+    test('serializes to JSON', () {
+      final user = User(
+          id: "user_id",
+          username: "username",
+          email: "email@email.com",
+          ipAddress: "127.0.0.1",
+          extras: {"foo": "bar"});
+
+      final breadcrumbs = [
+        Breadcrumb("test log", DateTime.utc(2019),
+            level: SeverityLevel.debug, category: "test"),
+      ];
+
+      expect(
+        Event(
+          message: 'test-message',
+          transaction: '/test/1',
+          exception: StateError('test-error'),
+          level: SeverityLevel.debug,
+          culprit: 'Professor Moriarty',
+          tags: <String, String>{
+            'a': 'b',
+            'c': 'd',
+          },
+          extra: <String, dynamic>{
+            'e': 'f',
+            'g': 2,
+          },
+          fingerprint: <String>[Event.defaultFingerprint, 'foo'],
+          userContext: user,
+          breadcrumbs: breadcrumbs,
+        ).toJson(),
+        <String, dynamic>{
+          'platform': 'dart',
+          'sdk': {'version': sdkVersion, 'name': 'dart'},
+          'message': 'test-message',
+          'transaction': '/test/1',
+          'exception': [
+            {'type': 'StateError', 'value': 'Bad state: test-error'}
+          ],
+          'level': 'debug',
+          'culprit': 'Professor Moriarty',
+          'tags': {'a': 'b', 'c': 'd'},
+          'extra': {'e': 'f', 'g': 2},
+          'fingerprint': ['{{ default }}', 'foo'],
+          'user': {
+            'id': 'user_id',
+            'username': 'username',
+            'email': 'email@email.com',
+            'ip_address': '127.0.0.1',
+            'extras': {'foo': 'bar'}
+          },
+          'breadcrumbs': {
+            'values': [
+              {
+                'timestamp': '2019-01-01T00:00:00',
+                'message': 'test log',
+                'category': 'test',
+                'level': 'debug',
+              },
+            ]
+          },
+        },
+      );
+    });
+  });
+}
+
+typedef Answer = dynamic Function(Invocation invocation);
+
+class MockClient implements Client {
+  Answer _answer;
+
+  void answerWith(Answer answer) {
+    _answer = answer;
+  }
+
+  noSuchMethod(Invocation invocation) {
+    return _answer(invocation);
+  }
 }