+++ test/version_test.dart
@@ -6,7 +6,7 @@
 
 import 'dart:io';
 
-import 'package:sentry/sentry.dart';
+import 'package:sentry/src/version.dart';
 import 'package:test/test.dart';
 import 'package:yaml/yaml.dart' as yaml;